+++ src/startup/index.ts
@@ -2,5 +2,5 @@ import * as dotenv from 'dotenv';
 
 export { dotenv };
 
-export { Startup } from './niflheim';
-export { Bootstrap, IBootstrapRoute } from './muspell';
+export { Startup } from './niflheim/startup';
+export { Bootstrap, IBootstrapRoute } from './muspell/bootstrap';
